drv:src: 14:aie4_aig - Merging DMA module from aie4_aig (#411)
This patch handles the merging of DMA module from aie4_aig

+++ driver/src/dma/xaie_dma_aie.c
@@ -1063,7 +1063,7 @@ AieRC _XAie_ShimDmaUpdateBdAddr(XAie_DevInst *DevInst,
 * @note		Internal only.
 *
 ******************************************************************************/
-AieRC _XAie_DmaSetBdIteration(XAie_DmaDesc *DmaDesc, u32 StepSize, u8 Wrap,
+AieRC _XAie_DmaSetBdIteration(XAie_DmaDesc *DmaDesc, u32 StepSize, u16 Wrap,
 		u8 IterCurr)
 {
 	(void)DmaDesc;
