+++ driver/src/dma/xaie_dma_aieml.h
@@ -69,7 +69,7 @@ AieRC _XAieMl_DmaUpdateBdAddr(XAie_DevInst *DevInst, const XAie_DmaMod *DmaMod,
 AieRC _XAieMl_ShimDmaUpdateBdAddr(XAie_DevInst *DevInst,
 		const XAie_DmaMod *DmaMod, XAie_LocType Loc, u64 Addr,
 		u16 BdNum);
-AieRC _XAieMl_DmaSetBdIteration(XAie_DmaDesc *DmaDesc, u32 StepSize, u8 Wrap,
+AieRC _XAieMl_DmaSetBdIteration(XAie_DmaDesc *DmaDesc, u32 StepSize, u16 Wrap,
 		u8 IterCurr);
 
 #endif /* XAIE_DMA_AIEML_H */
