+++ driver/src/global/xaiegbl_regdef.h
@@ -584,7 +584,7 @@ struct XAie_DmaMod {
 	AieRC (*SetIntrleave) (XAie_DmaDesc *Desc, u8 DoubleBuff,
 			u8 IntrleaveCount, u16 IntrleaveCurr);
 	AieRC (*SetMultiDim) (XAie_DmaDesc *Desc, XAie_DmaTensor *Tensor);
-	AieRC (*SetBdIter) (XAie_DmaDesc *Desc, u32 StepSize, u8 Wrap,
+	AieRC (*SetBdIter) (XAie_DmaDesc *Desc, u32 StepSize, u16 Wrap,
 			u8 IterCurr);
 	AieRC (*WriteBd)(XAie_DevInst *DevInst, XAie_DmaDesc *Desc,
 			XAie_LocType Loc, u16 BdNum);
